[fix unregister]

+++ drive/ressources/unregister.py
@@ -1,7 +1,7 @@
 from flask_restful import reqparse, Resource
 import json
 from flask import Response
-from models import App
+from models import App, Send_files
 from db import db
 
 parser = reqparse.RequestParser()
@@ -22,8 +22,19 @@ class Unregister(Resource):
                 response=json.dumps(dict(error='App does not exist')),
                 status=400, mimetype='application/json')
 
+        send = Send_files.query.filter_by(name_app_sending=args['app']).first()
+
+        if send is None:
+            return Response(
+                response=json.dumps(dict(info='App never sent something')),
+                status=200, mimetype='application/json')
+
+        db.session.delete(send)
+        db.session.commit()
+
         db.session.delete(app)
         db.session.commit()
+
 
         return Response(
             response=json.dumps(dict(info='you are successfully unregister')),
